Add files via upload
This project involved doing a client project under confidential terms and conditions. A patent was filed as an outcome of the work and the pdf is attached for reference. I have also attached pdf of a CAS drawing which was used to fabricate the belt vision system for this project.